Your terms of service are really bad
No offense to who wrote them, but  `the source code will be private only to the development team.` is really ambiguous: you wanted to say that the team *will* have access to the source, but you wrote it in such a way that it will be public for everybody else. Your lack of definitions (which I did not add) means that there is so much ambiguity. I've made a few changes, but it isn't enough. Also, you should add a footnote to every page that links to the ToSs.